Rename ConfigUpdater to ConfigReloader
